// frontend/generated/models/TransactionDto.ts
export interface TransactionDto {
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    transactionId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof TransactionDto
     */
    account?: AccountDto;
    /**
     * 
     * @type {string}
     * @memberof TransactionDto
     */
    type?: TransactionDtoTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof TransactionDto
     */
    paymentMethod?: TransactionDtoPaymentMethodEnum;
    /**
     * 
     * @type {Date}
     * @memberof TransactionDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof TransactionDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    amount?: number;
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    oldBalance?: number;
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    newBalance?: number;
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    orderId?: number;
    /**
     * 
     * @type {string}
     * @memberof TransactionDto
     */
    status?: TransactionDtoStatusEnum;
}


/**
 * @export
 */
export const TransactionDtoTypeEnum = {
    Deposit: 'DEPOSIT',
    Order: 'ORDER'
} as const;
export type TransactionDtoTypeEnum = typeof TransactionDtoTypeEnum[keyof typeof TransactionDtoTypeEnum];

/**
 * @export
 */
export const TransactionDtoPaymentMethodEnum = {
    InternalWallet: 'INTERNAL_WALLET',
    Paypal: 'PAYPAL',
    Vnpay: 'VNPAY'
} as const;
export type TransactionDtoPaymentMethodEnum = typeof TransactionDtoPaymentMethodEnum[keyof typeof TransactionDtoPaymentMethodEnum];

/**
 * @export
 */
export const TransactionDtoStatusEnum = {
    Pending: 'PENDING',
    Success: 'SUCCESS',
    Failed: 'FAILED'
} as const;
export type TransactionDtoStatusEnum = typeof TransactionDtoStatusEnum[keyof typeof TransactionDtoStatusEnum];