// frontend/generated/models/CreateDepositTransaction200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateDepositTransaction200Response
 */
export interface CreateDepositTransaction200Response {
    /**
     * 
     * @type {string}
     * @memberof CreateDepositTransaction200Response
     */
    paymentRedirectUrl?: string;
}

/**
 * Check if a given object implements the CreateDepositTransaction200Response interface.
 */
export function instanceOfCreateDepositTransaction200Response(value: object): value is CreateDepositTransaction200Response {
    return true;
}

export function CreateDepositTransaction200ResponseFromJSON(json: any): CreateDepositTransaction200Response {
    return CreateDepositTransaction200ResponseFromJSONTyped(json, false);
}

export function CreateDepositTransaction200ResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateDepositTransaction200Response {
    if (json == null) {
        return json;
    }
    return {
        
        'paymentRedirectUrl': json['paymentRedirectUrl'] == null ? undefined : json['paymentRedirectUrl'],
    };
}

export function CreateDepositTransaction200ResponseToJSON(json: any): CreateDepositTransaction200Response {
    return CreateDepositTransaction200ResponseToJSONTyped(json, false);
}

export function CreateDepositTransaction200ResponseToJSONTyped(value?: CreateDepositTransaction200Response | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'paymentRedirectUrl': value['paymentRedirectUrl'],
    };
}